feat: change price type from INTEGER to REAL in order and trade data structures

+++ agentpit_clob/trade.py
@@ -9,7 +9,7 @@ class Trade:
     maker_orders: list[dict[str, Any]]
     market: str
     asset_id: str
-    price: int
+    price: float
     trade_size: int
     remaining_size: int
     side: str
